HW-12: Fix clients scopes. Tweak SpringDocConfig. Tweak security.
- Tweak SpringDocConfig.
-  Clean.
- Tweak security.
- Config tweak.
- Add more method security.
