+++ hw12-spring-security/src/main/java/ru/otus/hw/api/rest/RestControllerExceptionHandler.java
@@ -27,7 +27,7 @@ public class RestControllerExceptionHandler extends ResponseEntityExceptionHandl
     }
 
     @ExceptionHandler(AuthorizationDeniedException.class)
-    public ResponseEntity<ProblemDetail> authorizationDeniedException(EntityNotFoundException ex) {
+    public ResponseEntity<ProblemDetail> authorizationDeniedException(AuthorizationDeniedException ex) {
         log.error(ex.getMessage(), ex);
         ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.FORBIDDEN,
             ex.getMessage());
