--- hw12-spring-security/src/main/resources/application-keycloak.yaml
@@ -1,20 +1,0 @@
-spring:
-  security:
-    oauth2:
-      client:
-        registration:
-          keycloak:
-            client-id: bookstore-client
-            client-secret: bookstore-client-secret
-            authorization-grant-type: authorization_code
-            redirect-uri: "${KEYCLOAK_BASE_URL:http://localhost:9090}/login/oauth2/code/keycloak"
-            scope: openid,profile,email
-        provider:
-          keycloak:
-            issuer-uri: "${KEYCLOAK_BASE_URL:http://localhost:58080}/realms/bookstore"
-            authorization-uri: "${KEYCLOAK_BASE_URL:http://localhost:58080}/realms/bookstore/protocol/openid-connect/auth"
-            token-uri: "${KEYCLOAK_BASE_URL:http://localhost:58080}/realms/bookstore/protocol/openid-connect/token"
-      resourceserver:
-        jwt:
-          issuer-uri: "${KEYCLOAK_BASE_URL:http://localhost:58080}/realms/bookstore"
-          jwk-set-uri: "${KEYCLOAK_BASE_URL:http://localhost:58080}/realms/bookstore/protocol/openid-connect/certs"
